requirements.txt
Django==3.0.2
django-tastypie==0.14.3
